requirements.txt
asgiref==3.5.2
Django==3.2.16
django-debug-toolbar==3.2.4
django-phonenumber-field==7.0.0
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.10.1
mccabe==0.7.0
phonenumbers==8.12.57
Pillow==9.2.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.4
sorl-thumbnail==12.9.0
sqlparse==0.4.3
typing_extensions==4.4.0
zipp==3.10.0
